requirements.txt
google-generativeai
python-dotenv
django
pillow
